code clean up and normalization
