+++ frontend/src/App.tsx
@@ -2,9 +2,7 @@ import './App.css'
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Login from "./pages/login/login";
 import Dashboard from "./pages/dashboard/dashboard";
-import FormPage from "./pages/form/form_page";
 import ProtectedRoute from './components/routeGuard/routeGuard';
-import ServicesPage from "./pages/services/services"
 import QueuePage from './pages/queue/queue';
 import AddPatient from './pages/add-patient/add_patient';
 import Records from './pages/patient-records/patient_records'; 
@@ -23,16 +21,6 @@ function App() {
         <Route path="/dashboard" element={
             <ProtectedRoute> 
               <Dashboard/> 
-            </ProtectedRoute>} 
-        />
-        <Route path="/forms" element={
-            <ProtectedRoute> 
-            <FormPage/>
-            </ProtectedRoute>} 
-        />
-        <Route path="/services" element={
-            <ProtectedRoute> 
-            <ServicesPage/>
             </ProtectedRoute>} 
         />
         <Route path="/queue" element={
